fix: address coach review feedback for feature/refactoring-and-ui PR - Replace fireEvent with userEvent in TournamentForm.test.js - Use getByRole('heading') instead of getByText in TournamentList.test.js - Remove double Link wrapper in Layout.js (StyledLink now extends Link) - Fix React key in TournamentCard (player name instead of array index) - Change Card from styled.li to styled.article; wrap in li in TournamentList - Extract processedRanking logic to pointsEngine.js as getProcessedRanking() - Remove unnecessary JSX comment in YearlyRanking.js
